Python_Project/requirements/requirements_2.txt
plotly==3.3.0
seaborn==0.9.0
beautifulsoup4==4.6.3
bs4==0.0.1
pyshp==1.2.10
geopandas==0.3.0
numpy==1.15.2
scipy==1.1.0
Shapely==1.6.3
requests==2.19.1